+++ modules/dnn/test/test_onnx_conformance_layer_parser_denylist.inl.hpp
@@ -93,12 +93,6 @@
 "test_det_nd",
 "test_dropout_default_mask",
 "test_dropout_default_mask_ratio",
-"test_dynamicquantizelinear",
-"test_dynamicquantizelinear_expanded",
-"test_dynamicquantizelinear_max_adjusted",
-"test_dynamicquantizelinear_max_adjusted_expanded",
-"test_dynamicquantizelinear_min_adjusted",
-"test_dynamicquantizelinear_min_adjusted_expanded",
 "test_edge_pad",
 "test_einsum_inner_prod",
 "test_elu_default_expanded_ver18",
